requirements.txt
python-telegram-bot[job_queue]==20.0
ccxt==4.4.75
requests==2.31.0
aiohttp==3.9.5
httpx==0.23.1
httpcore==0.15.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4